guarding routes & jwt on cokiees

+++ src/modules/recipes/recipes.controller.ts
@@ -7,10 +7,14 @@ import {
   Param,
   Patch,
   Post,
+  Request,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
 
 import { CreateRecipeDto, UpdateRecipeDto } from './dto/recipes.dto';
 import { RecipesService } from './recipes.service';
+import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
 
 @Controller('recipes')
 export class RecipesController {
@@ -27,21 +31,48 @@ export class RecipesController {
   }
 
   @Post()
-  async register(@Body() createRecipeDto: CreateRecipeDto): Promise<any> {
+  @UseGuards(JwtAuthGuard)
+  async register(
+    @Request() req: any,
+    @Body() createRecipeDto: CreateRecipeDto,
+  ): Promise<any> {
+    const id = req.user.id as string;
+
+    if (id !== createRecipeDto.authorId) {
+      throw new UnauthorizedException(
+        'You are not authorized to save a recipe on behalf of another user.',
+      );
+    }
+
     return await this.recipesService.register(createRecipeDto);
   }
 
   @Patch(':id')
+  @UseGuards(JwtAuthGuard)
   async update(
+    @Request() req: any,
     @Param('id') id: string,
     @Body() updateRecipeDto: UpdateRecipeDto,
   ): Promise<any> {
+    if (req.user.id !== id) {
+      throw new UnauthorizedException(
+        'You are not authorized to update a recipe on behalf of another user.',
+      );
+    }
+
     return await this.recipesService.update(id, updateRecipeDto);
   }
 
   @Delete(':id')
+  @UseGuards(JwtAuthGuard)
   @HttpCode(204)
-  async delete(@Param('id') id: string): Promise<any> {
+  async delete(@Request() req: any, @Param('id') id: string): Promise<any> {
+    if (req.user.id !== id) {
+      throw new UnauthorizedException(
+        'You are not authorized to delete a recipe on behalf of another user.',
+      );
+    }
+
     return await this.recipesService.delete(id);
   }
 }
